# api/imagebuilder/v1alpha1/openapi.yaml
openapi: 3.0.1
info:
  title: Flight Control Image Builder API
  version: v1alpha1
  description: |
    Image Builder service for Flight Control.
    This service manages image build resources for building and pushing container images.
  contact:
    name: The Flight Control Team
    url: https://flightctl.io
    email: team@flightctl.io
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: /
...
paths:
  /api/v1/imagebuilds:
    x-resource: imagebuilds
    get:
      tags:
        - imagebuild
      description: List ImageBuild resources.
      operationId: listImageBuilds
      parameters:
        - name: labelSelector
          in: query
          description: A selector to restrict the list of returned objects by their labels.
          schema:
            type: string
        - name: fieldSelector
          in: query
          description: A selector to restrict the list of returned objects by their fields.
          schema:
            type: string
        - name: limit
          in: query
          description: The maximum number of results returned in the list response.
          schema:
            type: integer
            format: int32
        - name: continue
          in: query
          description: An optional parameter to query more results from the server.
          schema:
            type: string
        - name: withExports
          in: query
          description: If true, includes related ImageExport resources in each ImageBuild in the response.
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageBuildList'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "429":
          description: Too Many Requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "503":
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
    post:
      tags:
        - imagebuild
      description: Create an ImageBuild resource.
      operationId: createImageBuild
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImageBuild'
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageBuild'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "429":
          description: Too Many Requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "503":
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
  /api/v1/imagebuilds/{name}:
    x-resource: imagebuilds
    get:
      tags:
        - imagebuild
      description: Get an ImageBuild resource.
      operationId: getImageBuild
      parameters:
        - name: name
          in: path
          description: The name of the ImageBuild resource.
          required: true
          schema:
            type: string
        - name: withExports
          in: query
          description: If true, includes related ImageExport resources in the response.
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageBuild'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "429":
          description: Too Many Requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "503":
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
    delete:
      tags:
        - imagebuild
      description: Delete an ImageBuild resource.
      operationId: deleteImageBuild
      parameters:
        - name: name
          in: path
          description: The name of the ImageBuild resource.
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageBuild'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "429":
          description: Too Many Requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "503":
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
  /api/v1/imagebuilds/{name}/log:
    x-resource: imagebuilds/log
    get:
      tags:
        - imagebuild
      description: Get logs for an ImageBuild resource. For active builds, streams logs live. For completed builds, returns last 500 lines.
      operationId: getImageBuildLog
      parameters:
        - name: name
          in: path
          description: The name of the ImageBuild resource.
          required: true
          schema:
            type: string
        - name: follow
          in: query
          description: If true, stream logs continuously (like kubectl logs -f). For active builds, keeps connection open. For completed builds, returns all logs and closes.
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: OK
          content:
            text/plain:
              schema:
                type: string
                description: Log output as plain text (when follow=false or for completed builds)
            text/event-stream:
              schema:
                type: string
                description: Log output as Server-Sent Events stream (when follow=true for active builds)
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "429":
          description: Too Many Requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "503":
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
  /api/v1/imagebuilds/{name}/cancel:
    x-resource: imagebuilds/cancel
    post:
      tags:
        - imagebuild
      description: Cancel a running ImageBuild. Only builds in Pending, Building, or Pushing state can be canceled.
      operationId: cancelImageBuild
      parameters:
        - name: name
          in: path
          description: The name of the ImageBuild resource to cancel.
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageBuild'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "409":
          description: Conflict - Build not in cancelable state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "429":
          description: Too Many Requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "503":
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
  /api/v1/imagedefinitions:
    x-resource: imagedefinitions
    get:
      tags:
        - imagedefinition
      description: List ImageDefinition resources.
      operationId: listImageDefinitions
      parameters:
        - name: labelSelector
          in: query
          description: A selector to restrict the list of returned objects by their labels.
          schema:
            type: string
        - name: fieldSelector
          in: query
          description: A selector to restrict the list of returned objects by their fields.
          schema:
            type: string
        - name: limit
          in: query
          description: The maximum number of results returned in the list response.
          schema:
            type: integer
            format: int32
        - name: continue
          in: query
          description: An optional parameter to query more results from the server.
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageDefinitionList'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "429":
          description: Too Many Requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "503":
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
    post:
      tags:
        - imagedefinition
      description: Create an ImageDefinition resource.
      operationId: createImageDefinition
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImageDefinition'
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageDefinition'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "429":
          description: Too Many Requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "503":
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
...
components:
  securitySchemes:
    bearerAuth:
      $ref: '../../core/v1beta1/openapi.yaml#/components/securitySchemes/bearerAuth'
    orgId:
      $ref: '../../core/v1beta1/openapi.yaml#/components/securitySchemes/orgId'
  schemas:
...
    ImageBuildList:
      type: object
      description: ImageBuildList is a list of ImageBuild resources.
      properties:
        apiVersion:
          $ref: '#/components/schemas/ApiVersion'
        kind:
          type: string
          description: 'Kind is a string value representing the REST resource this object represents. Servers may infer this from the endpoint the client submits requests to. Cannot be updated. In CamelCase. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds.'
        metadata:
          $ref: '../../core/v1beta1/openapi.yaml#/components/schemas/ListMeta'
        items:
          type: array
          description: 'List of ImageBuild resources.'
          items:
            $ref: '#/components/schemas/ImageBuild'
      required:
        - apiVersion
        - kind
        - metadata
        - items
...
    Status:
      type: object
      properties:
        apiVersion:
          $ref: '#/components/schemas/ApiVersion'
        kind:
          type: string
          description: 'Kind is a string value representing the REST resource this object represents. Servers may infer this from the endpoint the client submits requests to. Cannot be updated. In CamelCase. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds.'
        code:
          type: integer
          format: int32
          description: Suggested HTTP return code for this status, 0 if not set.
        message:
          type: string
          description: A human-readable description of the status of this operation.
        reason:
          type: string
          description: A machine-readable description of why this operation is in the "Failure" status. If this value is empty there is no information available. A Reason clarifies an HTTP status code but does not override it.
        status:
          type: string
          description: 'Status of the operation. One of: "Success" or "Failure". More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#spec-and-status.'
      description: Status is a return value for calls that don't return other objects.
      required:
        - apiVersion
        - kind
        - status
        - code
        - reason
        - message